+++ mondrian-charts/src/fiberplane/tests/test_generate_line_chart_from_timeseries.rs
@@ -47,6 +47,7 @@ fn test_generate_line_chart_from_timeseries() {
             grid_rows_shown: true,
             grid_stroke_color: "#e7e7e7",
             grid_stroke_dasharray: Default::default(),
+            shape_stroke_width: None,
             get_shape_list_color: &|_, _| "#c00eae",
             tick_color: "#a4a4a4",
             x_formatter: Some(FormatterKind::Time),